[BZ 185375] make sure TTL is set correctly and fix bug in TaskTracker
In some of my recent changes, I introduced a regression in MigrateData where
we were not handling the TTL properly. If the amount of time passed since the
data was writen is greater than the TTL, we do not insert it.

There was an scenario that TaskTracker did not account for. If all tasks
finished before finishedSchedulingTasks() is called, then the CountDownLatch is
never counted down, which means the caller of waitForTasksToFinish() would
block indefinitely.