[BZ 185375] do not delete partitions as migrations finish
When a metric migration finished, I had been deleting the partition from the
old table. The thinking was that it might help reduce heap pressure which in
turn would reduce GC pauses. I am not sure if doing the deletes helps with heap
pressure. I did not bother to investigate further because execution times have
been consistently faster without doing the deletes. I observed from the logs
that doing the deletes triggered compactions on the old tables. I attribute the
slow down to this.